+++ src/main/java/com/practise/categorymapping/service/ProductServiceImp.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.practise.categorymapping.exception.ProductNotFoundException;
 import com.practise.categorymapping.model.Product;
 import com.practise.categorymapping.repository.ProductRepository;
 
@@ -23,19 +24,26 @@ public class ProductServiceImp implements ProductService{
 
 	@Override
 	public Product viewProductInfo(int ProductId) {
-		return productRepository.findById(ProductId).orElseThrow(() -> new EntityNotFoundException("Product was not in the DATABASE"));
+		return productRepository.findById(ProductId).orElseThrow(() -> new ProductNotFoundException("Product was not in the DATABASE"));
 	}
 
 	@Override
 	public String newProduct(Product product) {
-		productRepository.save(product);
-		return "Product saved";
+		if(productRepository.save(product) != null) {
+			return "New Product Created";
+		}else {
+			throw new RuntimeException("Product was not saved in the database");
+		}
 	}
 
 	@Override
 	public String deleteProduct(int productId) {
-		productRepository.deleteById(productId);
-		return "Product Deleted";
+		if(productRepository.existsById(productId) == true) {
+			productRepository.deleteById(productId);
+			return "Product Deleted";
+		}else {
+			throw new RuntimeException("Product Id was not available in the database.");
+		}
 	}
 
 }
